Create gnome-pie-gdk-backends.vala
As Gnome-Pie will now run on the 'x11' GDK backend by default (in a Wayland session), we need to check whether that is available at runtime. Instead of only allowing the 'x11' GDK backend (Gdk.set_allowed_backends("x11");), dropping all wayland-specific workarounds and letting the app crash when the backend is not available, we opt to use a dedicated executable to check the availability of the 'x11' GDK backend at runtime. This needs a few extra shell calls via Posix.System("") - in daemon.vala - but is much cleaner and user-friendly this way.

IMPORTANT: due to the lack of CMake experience on my part, I might have placed this new file in the wrong location. It needs to be integrated into src/CMakeLists.txt, such that it is generated as a seperate executable (named gnome-pie-gdk-backends) but I don't know how to do that.